+++ templates/chats/chatroom.html
@@ -16,7 +16,7 @@
         <div id='chat_container' class="overflow-y-auto grow">
 					<ul id='chat_messages' class="flex flex-col justify-end gap-2 p-4">
 						{% for message in chat_messages reversed %}
-						{% include "chat/chatroom_msg.html" %}
+						{% include "chats/chatroom_msg.html" %}
 						{% endfor %}
 					</ul>
         </div>
