+++ templates/chats/chatroom_msg.html
@@ -13,7 +13,8 @@
 <li>
 	<div class="flex justify-start">
 		<div class="flex items-end mr-2" >
-			<a href="#">
+			<!-- 到時候下方的 a href 要連到會員資料頁面 -->
+			<a href="#">  
 				<img class="w-8 h-8 rounded-full object-cover" src="{{ message.author.profile.avatar }}">
 			</a>
 		</div>
