+++ templates/chats/chatroom_msg_partial.html
@@ -1,7 +1,7 @@
 <div id="chat_messages" hx-swap-oob="beforeend">
 
 	<div class="fade-in-up">
-		{% include "chat/chatroom_msg.html" %}
+		{% include "chats/chatroom_msg.html" %}
 	</div>
 
 	<style>
